requirements.txt
pydantic
tqdm
scikit-learn
llama-cpp-python
pandas
python-dotenv